Currently stuck
I'm trying to do a multi stage dockerfile build but it's failing to copy the go file over to the docker image. This is happening because I have no idea where the go file is or how to find it.

Also, you can't edit the sh file from windows or else it fails. If you do edit it, then you have to do a find and replace in notepad++. Replace \r\n with \n