Got goland and php to run on the same docker image!
I was able to get a 'hello world' app for both php and go running in the same docker image.

The piece I was missing before was that I wasn't copying over the necessary go files to run the go program.

The line that achieved this while creating the go image was:

COPY . .

This copies all of the go files over to the current image in the dockerfile, which after copying that image into the contents of the final image, I was able to run the go app as well as the php app, which was already working.

Here's a longer explanation for future reference:

To copy from multiple base images (php and go in this case) you have to first create one image, php or go with an alias, for example:

FROM golan:onbuild as gobase

The 'as gobase' is the alias bit. Once I've done that, I can create a second image, in this case the php image. Then, after wrapping up the php image, I can copy the go image to work alongside my new php image. This is done with the

COPY --from=gobase /build /