Got php server working outside of docker image
I got the php server working outside of docker. This can be done by running the command:

php -S localhost:(port)

This works as long as there is a index.php file setup in the current directory. When tested in postman I was able to successfully see the echo calls in the index file.

However I have been unable to send a request from my local machine to the php server when the server is inside the docker image. Not sure quite where to go with this one but it seems docker toolbox helps with this, though it does say that with windows 10 I don't need docker toolbox? Further investigation is needed.