Got the php server working.
I have now started using apache to run the server. I don't wholly understand the Apache part yet but I've got it mapping the port to my localhost, and sending http requests to my php file, so that's some progress.
Here is the link the blog I got the apache docker setup from: https://writing.pupius.co.uk/apache-and-php-on-docker-44faef716150

I'm currently working on being able to access the components of a POST request in php