Created the postback object
I created a php class for the postback object that will be passed from php to redis.

This class has the following structure:

class Postback {

    private $isValid;
    private $requestMethod;
    private $url;
    private $mascot;
    private $location;

Each data piece passed in from the incoming http request will be used to create a postback object, which will then be passed to Redis.