Changed the structure of the Postback object
Previously I had the data pieces of the postback object hardcoded to be Mascot and Location, giving a structure as follows:

  class Postback implements JsonSerializable {

    private $isValid;
    private $requestMethod;
    private $url;
    private $mascot;
    private $location;

This was to in order to match up with the example given in the project writeup.

I've changed the structure of this object to make the data aspect dynamic, so instead of mascot and location, it will simply store the data array and pass it on to Redis and Go. So as follows:

  class Postback implements JsonSerializable {

    private $isValid;
    private $requestMethod;
    private $url;
    private $dataArray;

I've made changes to the PHP code and the Go code to handle this.

Here is the structure of Postback in Go:

type Postback struct {
	RequestMethod string
	Url string
	Data []map[string]string
}

My next step is to make it so for each map in the Data variable, a http request is sent with the map fields inserted into their corresponding spot in the Url field.